# openapi.yaml
openapi: 3.0.1
info:
  title: Online Order API
  version: 1.0.0
  description: A basic API for working with the Swagger tools
paths:
  /orders:
    summary: Get all of the orders data
    description: >-
      This path is used to retrieve all of the orders data from the orders.json
      file.
    get:
      summary: Gets the order data
      description: Retrieve the order information from the orders.json file
      operationId: get_orders
      responses:
        '200':
          description: Retrieve orders
          content:
            application/json:
              examples:
                Orders:
                  value: |-
                    {
                      "orders": [
                        {
                          "name": "Carey Maynard",
                          "id": "001",
                          "state": "pending"
                        },
                        {
                          "name": "Angelo Ayala",
                          "id": "002",
                          "state": "canceled"
                        },
                        {
                          "name": "Regina Yates",
                          "id": "003",
                          "state": "pending"
                        },
                        {
                          "name": "Elliott Mcclure",
                          "id": "004",
                          "state": "pending"
                        }
                      ]
                    }
        '400':
          description: Failed orders retrieval
          content:
            application/json:
              examples:
                Failed Get:
                  value: >-
                    {"success": false, "message": "An unknown error has
                    occurred"}
      tags:
        - Orders
  /neworder:
    summary: Add a new order.
    description: This path is used to add a new order to the orders.json file.
    post:
      summary: Add a new order
      description: Add a new order to the orders.json file
      operationId: new_order
      requestBody:
        description: A new order object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Successfully create a new order
          content:
            application/json:
              examples:
                New Order:
                  value: |-
                    {
                      "name": "Valeria Carlson",
                      "id": "005",
                      "state": "pending"
                    }
      tags:
        - New Order
  /update/{id}:
    summary: Change the status of an order.
    description: >-
      This path is used to change the status of an order matching the provided
      id. It modifies the state attribute of each order.
    put:
      summary: Update the state of an order
      description: Update the state of an order with a matching id in the orders.json file
      operationId: update_order
      parameters:
        - name: id
          in: path
          description: 'The id of the order.'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              examples:
                Updated Order:
                  value: |-
                    {
                      "name": "Valeria Carlson",
                      "id": "005",
                      "state": "completed"
                    }
      tags:
        - Update Order
  /delete/{id}:
    summary: Delete an order
    description: This path is used to delete an order with a matching id
    delete:
      summary: Delete an order
      description: Delete an order with a matching id from the orders.json file
      operationId: delete_order
      parameters:
        - name: id
          in: path
          description: 'The id of the order.'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted order
          content:
            application/json:
              examples:
                Deleted:
                  value: '{"message": "Order successfully deleted"}'
      tags:
        - Delete Order
        
components:
  schemas:
    Order:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        state:
          type: string
      xml:
        name: Order